+++ AdMob/GoogleAdMobAdapterConfiguration.m
@@ -40,7 +40,7 @@ typedef NS_ENUM(NSInteger, AdMobAdapterErrorCode) {
 #pragma mark - MPAdapterConfiguration
 
 - (NSString *)adapterVersion {
-    return @"7.68.0.0";
+    return @"7.68.0.1";
 }
 
 - (NSString *)biddingToken {