+++ AdMob/MPGoogleAdMobInterstitialCustomEvent.m
@@ -132,6 +132,12 @@ didFailToReceiveAdWithError:(GADRequestError *)error {
 - (void)interstitialDidDismissScreen:(GADInterstitial *)ad {
     MPLogAdEvent([MPLogEvent adDidDisappearForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
     [self.delegate fullscreenAdAdapterAdDidDisappear:self];
+    
+    // Signal that the fullscreen ad is closing and the state should be reset.
+    // `fullscreenAdAdapterAdDidDismiss:` was introduced in MoPub SDK 5.15.0.
+    if ([self.delegate respondsToSelector:@selector(fullscreenAdAdapterAdDidDismiss:)]) {
+        [self.delegate fullscreenAdAdapterAdDidDismiss:self];
+    }
 }
 
 - (void)interstitialWillLeaveApplication:(GADInterstitial *)ad {